2025 Day 1
I spent too long trying to get the Part 2 division solution to work, when
brute force was easier to reason through and get to a working answer. I
really wanted to complete the division solution though, so I used Github
Copilot to get me there. It generated a few bogus answers using GPT 4.1,
but when I jumped to Gemini 2.5 Pro it got a working solution first try.
It was ugly though, so I spent some time refactoring the logic to read
cleaner.